코드 채점 요청시 DB 저장, 코드 채점 결과 조회, 스케줄러로 10초마다 자동으로 DB에서 결과가 나오지 않은 경우 결과 업데이트, 제출 기록 조회(최신 제출한 결과를 먼저)

+++ src/main/java/com/pocketco/domain/judge0/application/Judge0ServiceImpl.java
@@ -6,6 +6,16 @@ import com.pocketco.domain.judge0.converter.Judge0Converter;
 import com.pocketco.domain.judge0.dto.Judge0LanguageExternal;
 import com.pocketco.domain.judge0.dto.Judge0LanguageResponse;
 import com.pocketco.domain.judge0.dto.CodeSubmitRequest;
+import com.pocketco.domain.problem.entity.Problem;
+import com.pocketco.domain.user.entity.History;
+import com.pocketco.domain.user.entity.HistoryStatus;
+import com.pocketco.domain.user.entity.Judge0Token;
+import com.pocketco.domain.user.entity.User;
+import com.pocketco.domain.user.exception.UserNotFoundException;
+import com.pocketco.domain.user.repository.HistoryRepository;
+import com.pocketco.domain.user.exception.HistoryNotFoundException;
+import com.pocketco.domain.user.repository.Judge0TokenRepository;
+import com.pocketco.domain.user.repository.UserRepository;
 import com.pocketco.global.common.redis.RedisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,7 +23,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
 import com.pocketco.domain.judge0.dto.SubmissionResultResponse;
 
-import java.util.ArrayList;
 import java.util.UUID;
 import com.pocketco.domain.judge0.dto.*;
 import com.pocketco.domain.language.entity.Language;
@@ -36,6 +45,9 @@ public class Judge0ServiceImpl implements Judge0Service {
     private final RedisService redisService;
     private final LanguageRepository languageRepository;
     private final ProblemRepository problemRepository;
+    private final HistoryRepository historyRepository;
+    private final Judge0TokenRepository judge0TokenRepository;
+    private final UserRepository userRepository;
 
     @Override
     public List<Judge0LanguageResponse> getJudge0Languages() {
@@ -64,21 +76,47 @@ public class Judge0ServiceImpl implements Judge0Service {
 
 
     @Override
-    public String submitCode(Long problemId, String languageName, CodeSubmitRequest request) {
-        if (!problemRepository.existsById(problemId)) {
-            throw new ProblemHandler(ErrorStatus.PROBLEM_NOT_FOUND);
-        }
-
-        int languageId = languageRepository.findByName(languageName)
-                .map(Language::getCode)
-                .orElseThrow(LanguageNotFoundException::new);
+    public SubmissionResponse submitCode(Long userId, Long problemId, String languageName, CodeSubmitRequest request) {
+        User me = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
+        Problem problem = problemRepository.findById(problemId).orElseThrow(() -> new ProblemHandler(ErrorStatus.PROBLEM_NOT_FOUND));
+        Language language = languageRepository.findByName(languageName).orElseThrow(LanguageNotFoundException::new);
+        int languageId = language.getCode();
 
         List<String> realTokens = fetchRealTokensFromJudge0(problemId, languageId, request, false);
 
         String submissionId = UUID.randomUUID().toString();
         redisService.saveTokens(submissionId, realTokens);
 
-        return submissionId;
+        // 단 하나라도 정답을 맞춘 적이 있다면 isSolved == true 값 저장
+        boolean isSolved = historyRepository.existsByUser_IdAndProblem_IdAndStatus(userId, problemId, HistoryStatus.ACCEPTED);
+
+        // DB 저장
+        History newHistory = History.builder()
+                .sourceCode(request.sourceCode())
+                .status(HistoryStatus.PROCESSING)
+                .isSolved(isSolved)
+                .isSolutionViewed(false)
+                .isGoalMet(false)
+                .user(me)
+                .language(language)
+                .problem(problem)
+                .build();
+
+        History savedHistory = historyRepository.save(newHistory);
+
+        for (String token : realTokens) {
+            Judge0Token judge0Token = Judge0Token.builder()
+                    .token(token)
+                    .statusId(1)
+                    .history(savedHistory)
+                    .build();
+            judge0TokenRepository.save(judge0Token);
+        }
+
+        return SubmissionResponse.builder()
+                .historyId(savedHistory.getId())
+                .submissionId(submissionId)
+                .build();
     }
 
 
@@ -140,40 +178,56 @@ public class Judge0ServiceImpl implements Judge0Service {
     }
 
     @Override
-    public SubmissionResultResponse getResult(String submissionId) {
-        List<String> tokens = redisService.getTokens(submissionId);
-
-        for (String token : tokens) {
-            Judge0StatusResponse response = webClient.get()
-                    .uri(uriBuilder -> uriBuilder
-                            .path("/submissions/{token}")
-                            .queryParam("fields", "status")
-                            .build(token))
-                    .retrieve()
-                    .bodyToMono(Judge0StatusResponse.class)
-                    .block();
-
-            if (response != null) {
-                int statusId = response.status().id();
-
-                if (statusId == 1 || statusId == 2) {
-                    return SubmissionResultResponse.builder()
-                            .status("PROCESSING")
-                            .message("채점중입니다")
-                            .build();
-                }
-            }
+    public SubmissionResultResponse getSubmitResult(Long historyId, String submissionId) {
+        History history = historyRepository.findById(historyId).orElseThrow(HistoryNotFoundException::new);
+        // 이미 채점이 완료되어 DB에 저장되어 있으면 채점이 완료 된 상태로 반환
+        if (history.getStatus() != HistoryStatus.PROCESSING) {
+            return SubmissionResultResponse.builder()
+                    .success(true)
+                    .status(history.getStatus())
+                    .message("채점이 이미 완료되었습니다.")
+                    .build();
         }
 
-        redisService.deleteTokens(submissionId);
+        List<String> tokens = redisService.getTokens(submissionId);
+        // redis에서 token을 못가져왔을 때
+        if (tokens == null || tokens.isEmpty()) {
+            return SubmissionResultResponse.builder()
+                    .success(false)
+                    .status(history.getStatus())
+                    .message("submissionId로 토큰을 얻어오지 못했습니다.")
+                    .build();
+        }
+
+        Judge0ResultResponse response = getJudge0ResultStatus(tokens);
+
+        boolean allDone = response.allDone();
+        HistoryStatus newStatus = response.newStatus();
+
+        if (!allDone) {
+            return SubmissionResultResponse.builder()
+                    .success(false)
+                    .status(HistoryStatus.PROCESSING)
+                    .message("채점중입니다...")
+                    .build();
+        }
+
+        int updated = historyRepository.updateStatus(historyId, newStatus);
+
+        // 스케줄러에 의해 이미 업데이트 된 경우
+        if (updated == 0) {
+            newStatus = historyRepository.findById(historyId).orElseThrow(HistoryNotFoundException::new).getStatus();
+        }
+        if (updated == 1) {
+            redisService.deleteTokens(submissionId);
+        }
 
         return SubmissionResultResponse.builder()
-                .status("DONE")
+                .success(true)
+                .status(newStatus)
                 .message("채점이 완료되었습니다.")
                 .build();
     }
-
-
 
     private List<String> fetchRealTokensFromJudge0(Long problemId, int languageId, CodeSubmitRequest request, boolean isSampleOnly) {
         //String encodedSource = Base64.getEncoder()
@@ -210,5 +264,80 @@ public class Judge0ServiceImpl implements Judge0Service {
                 .stream()
                 .map(Judge0TokenResponse::token)
                 .toList();
+    }
+
+    @Override
+    public Judge0ResultResponse getJudge0ResultStatus(List<String> tokens) {
+        // DB 상태 업데이트를 위한 메서드
+
+        HistoryStatus newStatus = HistoryStatus.PROCESSING;
+        boolean allDone = true;
+
+        for (String token : tokens) {
+            try {
+                Judge0StatusResponse response = webClient.get()
+                        .uri(uriBuilder -> uriBuilder
+                                .path("/submissions/{token}")
+                                .queryParam("fields", "status")
+                                .build(token))
+                        .retrieve()
+                        .bodyToMono(Judge0StatusResponse.class)
+                        .block();
+
+                if (response == null || response.status() == null) {
+                    allDone = false;
+                    continue;
+                }
+                int statusId = response.status().id();
+
+                if (statusId == 1 || statusId == 2) {
+                    allDone = false;
+                    continue;
+                }
+                newStatus = calcStatus(newStatus, statusId);
+                judge0TokenRepository.updateStatusWithToken(token, statusId);
+            } catch (Exception e) {
+                allDone = false;
+            }
+        }
+
+        return Judge0ResultResponse.builder()
+                .allDone(allDone)
+                .newStatus(newStatus)
+                .build();
+    }
+
+    private HistoryStatus calcStatus(HistoryStatus nowStatus, int statusId) {
+        HistoryStatus status = nowStatus;
+
+        // 상태 업데이트 우선 순위 => Compilation Error > System Error == Runtime Error > Time Limit Exceeded > Wrong Answer > Accepted
+        if (statusId == 3 && nowStatus == HistoryStatus.PROCESSING) {
+            status = HistoryStatus.ACCEPTED;
+        }
+        if (statusId == 4) {
+            if (nowStatus == HistoryStatus.PROCESSING
+                    || nowStatus== HistoryStatus.ACCEPTED) {
+                status = HistoryStatus.WRONG_ANSWER;
+            }
+        }
+        if (statusId == 5) {
+            if (nowStatus != HistoryStatus.RUNTIME_ERROR
+                    && nowStatus != HistoryStatus.COMPILATION_ERROR) {
+                status = HistoryStatus.TIME_LIMIT_EXCEEDED;
+            }
+        }
+        if (statusId == 6) {
+            status = HistoryStatus.COMPILATION_ERROR;
+        }
+        if (statusId > 6) {
+            if (nowStatus != HistoryStatus.COMPILATION_ERROR) {
+                status = HistoryStatus.RUNTIME_ERROR;
+            }
+            if (statusId == 13) {
+                status = HistoryStatus.SYSTEM_ERROR;
+            }
+        }
+
+        return status;
     }
 }
